+++ plugins/auth-supertokens/src/static/admin-ui/emailpassword/LoginForm.tsx
@@ -1,0 +1,46 @@
+import * as React from "react";
+import { useState } from "react";
+import { useLogin, useNotify } from "react-admin";
+import { Button } from "@mui/material";
+
+
+const LoginForm = ({ theme }: any) => {
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const login = useLogin();
+  const notify = useNotify();
+  const submit = async (e: any) => {
+    e.preventDefault();
+    login({ email, password }).catch(() => notify("Invalid email or password"));
+  };
+
+  return (
+            <form onSubmit={submit}>
+              <label>
+                <span>Email</span>
+
+                <input
+                  name="email"
+                  type="email"
+                  value={email}
+                  onChange={(e) => setEmail(e.target.value)}
+                />
+              </label>
+              <label>
+                <span>Password</span>
+
+                <input
+                  name="password"
+                  type="password"
+                  value={password}
+                  onChange={(e) => setPassword(e.target.value)}
+                />
+              </label>
+              <Button type="submit" variant="contained" color="primary">
+                Log in
+              </Button>
+            </form>
+  );
+};
+
+export default LoginForm;
