+++ plugins/auth-supertokens/src/templates/admin-ui/thirdpartyemailpassword/LoginForm.tsx
@@ -1,41 +1,24 @@
 import * as React from "react";
 import { useState } from "react";
 import { useLogin, useNotify, Notification, defaultTheme } from "react-admin";
-import { ThemeProvider } from "@material-ui/styles";
-import { createTheme } from "@material-ui/core/styles";
-import { Button } from "@material-ui/core";
+import { Button } from "@mui/material";
 import { getAuthorisationURLWithQueryParamsAndSetState } from "supertokens-web-js/recipe/thirdparty";
 import { SuperTokensConfig } from "./config";
-import "./login.scss";
 
-const CLASS_NAME = "login-page";
 
-const Login = ({ theme }: any) => {
+const LoginForm = ({ theme }: any) => {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const login = useLogin();
   const notify = useNotify();
-  const BASE_URI = process.env.REACT_APP_SERVER_URL;
   const submit = async (e: any) => {
     e.preventDefault();
     login({ email, password }).catch(() => notify("Invalid email or password"));
   };
 
   return (
-    <ThemeProvider theme={createTheme(defaultTheme)}>
-      <div className={`${CLASS_NAME}`}>
-        <div className={`${CLASS_NAME}__wrapper`}>
-          <div className={`${CLASS_NAME}__box`}>
-            <img
-              src="https://amplication.com/assets/react-admin.png"
-              alt="React-Admin"
-            />
-            <h2>Admin UI</h2>
-            <div className={`${CLASS_NAME}__box__message`}>
-              Sign in to a React-Admin client with ready-made forms for creating
-              and editing all the data models of your application
-            </div>
-            {SIGN_IN_BUTTONS}
+<>
+{SIGN_IN_BUTTONS}
             <p>or</p>
             <form onSubmit={submit}>
               <label>
@@ -62,38 +45,7 @@ const Login = ({ theme }: any) => {
                 Log in
               </Button>
             </form>
-          </div>
-          <div className={`${CLASS_NAME}__box`}>
-            <img
-              src="https://amplication.com/assets/restapi.png"
-              alt="REST API"
-            />
-            <h2>Connect via REST API</h2>
-            <div className={`${CLASS_NAME}__box__message`}>
-              Connect to the server using REST API with a built-in Swagger
-              documentation
-            </div>
-            <Button
-              type="button"
-              variant="contained"
-              color="primary"
-              href={`${BASE_URI}/api`}
-            >
-              Continue
-            </Button>
-          </div>
-
-          <Notification />
-        </div>
-        <div className={`${CLASS_NAME}__read-more`}>
-          <span>Read </span>
-          <a href="https://docs.amplication.com/api" target="docs">
-            Amplication docs
-          </a>
-          <span> to learn more</span>
-        </div>
-      </div>
-    </ThemeProvider>
+</>
   );
 };
 
@@ -129,4 +81,4 @@ const signInClicked = (thirdPartyId: string) => {
   };
 };
 
-export default Login;
+export default LoginForm;
